Move POJO model/mapping contributors and collectors to SPIs to allow for mapper-specific contributions
