+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/extractor/impl/BoundContainerValueExtractorPath.java
@@ -6,6 +6,7 @@
  */
 package org.hibernate.search.v6poc.entity.pojo.extractor.impl;
 
+import org.hibernate.search.v6poc.entity.pojo.extractor.spi.ContainerValueExtractorPath;
 import org.hibernate.search.v6poc.entity.pojo.model.spi.PojoGenericTypeModel;
 
 /**
