+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/mapping/building/spi/PojoMappingCollectorValueNode.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.entity.pojo.mapping.building.impl;
+package org.hibernate.search.v6poc.entity.pojo.mapping.building.spi;
 
 import java.util.Set;
