+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/mapping/definition/programmatic/impl/AssociationInverseSideMappingContextImpl.java
@@ -9,12 +9,12 @@ package org.hibernate.search.v6poc.entity.pojo.mapping.definition.programmatic.i
 import java.util.List;
 
 import org.hibernate.search.v6poc.entity.pojo.extractor.ContainerValueExtractor;
-import org.hibernate.search.v6poc.entity.pojo.extractor.impl.ContainerValueExtractorPath;
-import org.hibernate.search.v6poc.entity.pojo.mapping.building.impl.PojoMappingCollector;
-import org.hibernate.search.v6poc.entity.pojo.mapping.building.impl.PojoMetadataContributor;
+import org.hibernate.search.v6poc.entity.pojo.extractor.spi.ContainerValueExtractorPath;
+import org.hibernate.search.v6poc.entity.pojo.mapping.building.spi.PojoMappingCollector;
+import org.hibernate.search.v6poc.entity.pojo.mapping.building.spi.PojoMetadataContributor;
 import org.hibernate.search.v6poc.entity.pojo.mapping.definition.programmatic.AssociationInverseSideMappingContext;
 import org.hibernate.search.v6poc.entity.pojo.mapping.definition.programmatic.PropertyMappingContext;
-import org.hibernate.search.v6poc.entity.pojo.model.augmented.building.impl.PojoAugmentedModelCollectorPropertyNode;
+import org.hibernate.search.v6poc.entity.pojo.model.augmented.building.spi.PojoAugmentedModelCollectorPropertyNode;
 
 
 /**
