+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/model/augmented/building/impl/PojoAugmentedValueModelBuilder.java
@@ -6,7 +6,8 @@
  */
 package org.hibernate.search.v6poc.entity.pojo.model.augmented.building.impl;
 
-import org.hibernate.search.v6poc.entity.pojo.extractor.impl.ContainerValueExtractorPath;
+import org.hibernate.search.v6poc.entity.pojo.extractor.spi.ContainerValueExtractorPath;
+import org.hibernate.search.v6poc.entity.pojo.model.augmented.building.spi.PojoAugmentedModelCollectorValueNode;
 import org.hibernate.search.v6poc.entity.pojo.model.augmented.impl.PojoAugmentedValueModel;
 import org.hibernate.search.v6poc.entity.pojo.model.augmented.impl.PojoAssociationPath;
