+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/model/augmented/building/spi/PojoAugmentedModelCollectorValueNode.java
@@ -4,9 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.entity.pojo.model.augmented.building.impl;
+package org.hibernate.search.v6poc.entity.pojo.model.augmented.building.spi;
 
-import org.hibernate.search.v6poc.entity.pojo.extractor.impl.ContainerValueExtractorPath;
+import org.hibernate.search.v6poc.entity.pojo.extractor.spi.ContainerValueExtractorPath;
 
 public interface PojoAugmentedModelCollectorValueNode extends PojoAugmentedModelCollector {
